+++ compiler-compat/k240_dev_2124/src/main/kotlin/dev/zacsweers/metro/compiler/compat/k240_dev_2124/CompatContextImpl.kt
@@ -19,7 +19,6 @@ import org.jetbrains.kotlin.fir.expressions.FirExpressionEvaluator
 import org.jetbrains.kotlin.fir.expressions.PrivateConstantEvaluatorAPI
 import org.jetbrains.kotlin.fir.extensions.FirExtensionRegistrar
 import org.jetbrains.kotlin.fir.extensions.FirExtensionRegistrarAdapter
-import org.jetbrains.kotlin.fir.unwrapOr
 import org.jetbrains.kotlin.ir.builders.IrBuilder
 import org.jetbrains.kotlin.ir.builders.irAnnotation
 import org.jetbrains.kotlin.ir.expressions.IrConstructorCall
@@ -64,21 +63,19 @@ public class CompatContextImpl : CompatContext by DelegateType() {
     return FirExpressionEvaluator.evaluateExpression(this, session)?.unwrapOr {}
   }
 
-  private fun <T : FirElement> FirEvaluatorResult.unwrapOr(
-    action: (CompileTimeException) -> Unit
-  ): T? {
-    @Suppress("UNCHECKED_CAST")
-    when (this) {
-      is CompileTimeException -> action(this)
-      is Evaluated -> return this.result as? T
-      else -> return null
-    }
-    return null
-  }
-
   public class Factory : CompatContext.Factory {
     override val minVersion: String = "2.4.0-dev-2124"
 
     override fun create(): CompatContext = CompatContextImpl()
   }
+}
+
+fun <T : FirElement> FirEvaluatorResult.unwrapOr(action: (CompileTimeException) -> Unit): T? {
+  @Suppress("UNCHECKED_CAST")
+  when (this) {
+    is CompileTimeException -> action(this)
+    is Evaluated -> return this.result as? T
+    else -> return null
+  }
+  return null
 }