+++ compiler-tests/src/test/data/diagnostic/dependencygraph/dynamic/DuplicateBindingsInDynamicGraphReportTheCallLocation.kt
@@ -15,5 +15,5 @@ object TestBindings {
 }
 
 fun example() {
-  <!METRO_ERROR!>createDynamicGraph<AppGraph>(TestBindings)<!>
+  <!DUPLICATE_BINDING!>createDynamicGraph<AppGraph>(TestBindings)<!>
 }